requirements.txt
Django==4.2.7
Pillow==10.1.0
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
whitenoise==6.6.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-jazzmin==2.6.0 